# backend/swagger.yaml
swagger: "2.0"
info:
  description: "A web-based  judging  application  that  allows  users,  namely Admin, to host events and Participants, to create their project ideas or join an existing project and Judges, to be seamlessly involved in rating the assigned projects. Participants create and join the projects in an event. Eventually, judges are allowed to rate and admins can announce the winners of the projects in each event that are assigned to them. Each user has their roles and access to perform their respective operations described in the upcoming sections."
  version: "1.0.0"
  title: "SeamlessJudging"
host: "stark.cse.buffalo:3000"
schemes:
- "http"
paths:
...
  /api/v1/sponsor:
    get:
      tags:
      - "Sponsor"
      summary: "Retrieve Sponsor Details"
      description: "Only admin can retreive sponsors list"
      operationId: "getSponsors"
      produces:
      - "application/json"
      parameters:
      - name: "Request-ID"
        in: "header"
        required: true
        type: "string"
      - name: "Access-Token"
        in: "header"
        description: ""
        required: true
        type: "string"
      responses:
          "200":
            description: "successful"
            schema:
              type: "object"
              properties:
                response_str:
                    type: "string"
                    example: "Sponsor details retrieved successfully"
                response_data:
                  type: "array"
                  items:
                    $ref: "#/definitions/Sponsor"
          "404":
            description: "Not Found"

    post:
      tags:
      - "Sponsor"
      summary: "Add a Sponsor"
      description: "Only admin can add a sponsor"
      operationId: "addSponsor"
      produces:
      - "application/json"
      parameters:
      - name: "Request-ID"
        in: "header"
        required: true
        type: "string"
      - name: "Access-Token"
        in: "header"
        description: ""
        required: true
        type: "string"
      - in: "body"
        name: "body"
        description: "Request to add a sponsor"
        required: true
        schema:
          $ref: "#/definitions/AddSponsorRequest"
      responses:
          "201":
            description: "successfully added"
            schema:
              type: "object"
              properties:
                response_str:
                    type: "string"
                    example: "Sponsor added succesfully"
                response_data:
                  type: "object"
          "404":
            description: "Not Found"
    
    delete:
      tags:
      - "Sponsor"
      summary: "Delete sponsors in bulk"
      description: "Delete sponsors in bulk"
      operationId: "deleteSponsors"
      produces:
      - "application/json"
      parameters:
      - name: "Access-Token"
        in: "header"
        description: ""
        required: true
        type: "string"
      - in: "body"
        name: "body"
        description: "Request to delete sponsors"
        required: true
        schema:
          type: "object"
          properties:
            ids:
              $ref: "#/definitions/ArrayOfInt"
      responses:
          "201":
            description: "Sponsors deleted successfully"
            schema:
              type: "object"
              properties:
                response_str:
                    type: "string"
                    example: "Sponsors deleted successfully"
                response_data:
                  type: "object"
          "404":
            description: "Not Found"
  /api/v1/sponsor/{sponsorId}:
    put:
      tags:
      - "Sponsor"
      summary: "Update a Sponsor detail"
      description: "Only admin can update a sponsor"
      operationId: "updateSponsor"
      produces:
      - "application/json"
      parameters:
      - name: "Access-Token"
        in: "header"
        description: ""
        required: true
        type: "string"
      - name: "sponsorId"
        in: "path"
        description: ""
        required: true
        type: "string"
      - in: "body"
        name: "body"
        description: "Request to update a judge"
        required: true
        schema:
          $ref: "#/definitions/AddSponsorRequest"
      responses:
          "200":
            description: "successfully updated"
            schema:
              type: "object"
              properties:
                response_str:
                    type: "string"
                    example: "Sponsor updated succesfully"
                response_data:
                  type: "object"
          "404":
            description: "Not Found"
...
  /api/v1/winner-categories:
    get:
      tags:
      - "WinnerCategories"
      summary: "Retrieve Winner Categories Details"
      description: "Only admin retreive winner-categories complete list"
      operationId: "getWinnerCategories"
      produces:
      - "application/json"
      parameters:
      - name: "Access-Token"
        in: "header"
        description: ""
        required: true
        type: "string"
      responses:
          "200":
            description: "successful"
            schema:
              type: "object"
              properties:
                response_str:
                    type: "string"
                    example: "Winner Categories details retrieved successfully"
                response_data:
                  type: "array"
                  items:
                    $ref: "#/definitions/WinnerCategory"
          "404":
            description: "Not Found"

    post:
      tags:
      - "WinnerCategories"
      summary: "Add a WinnerCategory"
      description: "Only admin can add a WinnerCategory"
      operationId: "addWinnerCategory"
      produces:
      - "application/json"
      parameters:
      - name: "Request-ID"
        in: "header"
        required: true
        type: "string"
      - name: "Access-Token"
        in: "header"
        description: ""
        required: true
        type: "string"
      - in: "body"
        name: "body"
        description: "Add a WinnerCategory"
        required: true
        schema:
          type: "object"
          properties:
            name:
              type: "string"
              example: "First Prize"
      responses:
          "201":
            description: "successfully added"
            schema:
              type: "object"
              properties:
                response_str:
                    type: "string"
                    example: "WinnerCategory added succesfully"
                response_data:
                  type: "object"
          "404":
            description: "Not Found"
    
    delete:
      tags:
      - "WinnerCategories"
      summary: "Delete winner-categories in bulk"
      description: "Delete winner-categories in bulk"
      operationId: "deleteWinnerCategories"
      produces:
      - "application/json"
      parameters:
      - name: "Access-Token"
        in: "header"
        description: ""
        required: true
        type: "string"
      - in: "body"
        name: "body"
        description: "Request to delete winner-categories"
        required: true
        schema:
          type: "object"
          properties:
            ids:
              $ref: "#/definitions/ArrayOfInt"
      responses:
          "201":
            description: "Winner-categories deleted successfully"
            schema:
              type: "object"
              properties:
                response_str:
                    type: "string"
                    example: "Winner-categories deleted successfully"
                response_data:
                  type: "object"
          "404":
            description: "Not Found"
  /api/v1/winner-categories/{winnerCategoryId}:
    put:
      tags:
      - "WinnerCategories"
      summary: "Update a WinnerCategory"
      description: "Only admin can update a WinnerCategory"
      operationId: "updateWinnerCategory"
      produces:
      - "application/json"
      parameters:
      - name: "Access-Token"
        in: "header"
        description: ""
        required: true
        type: "string"
      - name: "winnerCategoryId"
        in: "path"
        description: ""
        required: true
        type: "string"
      - in: "body"
        name: "body"
        description: "Request to update a WinnerCategory"
        required: true
        schema:
          type: "object"
          properties:
            name:
              type: "string"
              example: "First Prize"
      responses:
          "200":
            description: "successfully updated"
            schema:
              type: "object"
              properties:
                response_str:
                    type: "string"
                    example: "WinnerCategory updated succesfully"
                response_data:
                  type: "object"
          "404":
            description: "Not Found"
...
definitions:
...
  ArrayOfInt:
    type: array
    items:
      type: integer
...
  Sponsor:
    type: "object"
    properties:
      sponsor_id:
          type: integer
          example: 1
      name:
          type: "string"
          example: "University at Buffalo"
      logo:
          type: "string"
          example: "logo.png"
  AddSponsorRequest:
    type: "object"
    properties:
      name:
          type: "string"
          example: "University at Buffalo"
      logo:
          type: "string"
          example: "logo.png"
  WinnerCategory:
    type: "object"
    properties:
      winner_category_id:
          type: integer
          example: 1
      name:
          type: "string"
          example: "First Prize"
...